Created cmor/_utils.py
