+++ esmvalcore/cmor/check.py
@@ -15,14 +15,14 @@ import iris.util
 import numpy as np
 from iris.cube import Cube
 
-from esmvalcore.cmor import (
+from esmvalcore.cmor._fixes.fix import AutomaticFix
+from esmvalcore.cmor._utils import (
     _get_alternative_generic_lev_coord,
     _get_generic_lev_coord_names,
     _get_new_generic_level_coord,
     _get_simplified_calendar,
     _is_unstructured_grid,
 )
-from esmvalcore.cmor._fixes.fix import AutomaticFix
 from esmvalcore.cmor.table import get_var_info
 from esmvalcore.exceptions import ESMValCoreDeprecationWarning
